Refactor workflow_sycn api to have everything in the 'service'
